+++ src/components/Tutorial/Tutorial.tsx
@@ -2,10 +2,11 @@ import React, { useEffect, useState } from 'react';
 import ReactDOM from 'react-dom';
 import styles from './Tutorial.module.scss';
 import classnames from 'classnames';
-import { ComponentChildren } from 'App.types';
+import { ComponentChildren, Unshaped } from 'App.types';
 import { ReactComponent as ArrowIcon } from 'assets/icons/Arrow__primary-white.svg';
 import { usePhoneScreen } from 'hooks/media.hook';
 import { throttle } from 'throttle-debounce';
+import useModalRoot from '../../hooks/modalRoot.hook';
 
 export interface ComponentProps {
   children?: ComponentChildren;
@@ -22,7 +23,14 @@ function calcPos() {
     'MyAreaHeader',
     'MyTipsHeader',
   ].reduce((acc, id) => {
-    const rect = document.getElementById(id)!.getBoundingClientRect();
+    const rectEl = document.getElementById(id);
+    let rect = {
+      top: 0,
+      left: 0,
+    };
+    if (rectEl && rectEl.getBoundingClientRect) {
+      rect = rectEl.getBoundingClientRect();
+    }
     return {
       ...acc,
       [id]: {
@@ -34,7 +42,7 @@ function calcPos() {
 }
 
 export default function Tutorial({ toggleTutorial }: ComponentProps) {
-  const [pos, setPos]: [any, any] = useState({});
+  const [pos, setPos]: [Unshaped, (pos: Unshaped) => void] = useState({});
 
   function handleEscapeKey(e: KeyboardEvent) {
     if (e.keyCode === 27) {
@@ -63,10 +71,7 @@ export default function Tutorial({ toggleTutorial }: ComponentProps) {
   // Check if positions are calculated
   return pos.MyUpdatesHeader ? (
     ReactDOM.createPortal(
-      <div
-        className={styles.Tutorial}
-        style={{ height: document.body.clientHeight }}
-      >
+      <div className={styles.Tutorial}>
         <div
           className={classnames(styles.TutorialItem, styles.MyUpdatesItem)}
           style={{
@@ -150,7 +155,7 @@ export default function Tutorial({ toggleTutorial }: ComponentProps) {
           voorzieningen van de gemeente
         </div>
       </div>,
-      document.getElementById('modal-root')!
+      useModalRoot()
     )
   ) : (
     <></>
